otras cosas

+++ CultivosCR/routes/web.php
@@ -28,26 +28,27 @@ Route::get('/favoriteGardens', 'UserController@favoriteGardens');
 Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
-Route::get('admin/garden/{id}/Sales', 'GardenController@sales')->name('sales');
-Route::get('admin/garden/{id}/Sales/create', 'GardenController@createSale')->name('sales/create');
-Route::post('admin/garden/{id}/Sales/create', 'GardenController@insertSale')->name('sales/create');
-Route::get('admin/garden/{id}/Sales/{idSale}', 'GardenController@saledetail');
-Route::get('admin/garden/{id}/Trades', 'GardenController@trades')->name('trades');
-Route::get('admin/garden/{id}/Trades/{idTrade}', 'GardenController@tradedetail');
-Route::get('admin/garden/{id}/Trades/create', 'GardenController@createTrade')->name('trades/create');
+Route::get('admin/garden/{id}/Sales', 'GardenController@sales')->name('sales')->middleware('auth');
+Route::get('admin/garden/{id}/Sales/create', 'GardenController@createSale')->name('sales/create')->middleware('auth');
+Route::post('admin/garden/{id}/Sales/create', 'GardenController@insertSale')->name('sales/create')->middleware('auth');
+Route::get('admin/garden/{id}/Sales/{idSale}', 'GardenController@saledetail')->middleware('auth');
+Route::get('admin/garden/{id}/Trades', 'GardenController@trades')->name('trades')->middleware('auth');
+Route::get('admin/garden/{id}/Trades/create', 'GardenController@createTrade')->name('trades/create')->middleware('auth');
+Route::get('admin/garden/{id}/Trades/{idTrade}', 'GardenController@tradedetail')->middleware('auth');
 
-Route::get('admin/garden/{id}', 'GardenController@showAdmin');
-Route::get('admin/garden/{id}/products', 'GardenController@productsAdmin');
-Route::post('admin/garden/{id}', 'GardenController@storeReview')->name('gardens.review');
-Route::delete('admin/garden/{id}', 'GardenController@destroyReview')->name('gardens.review.destroy');
+
+Route::get('admin/garden/{id}', 'GardenController@showAdmin')->middleware('auth');
+Route::get('admin/garden/{id}/products', 'GardenController@productsAdmin')->middleware('auth');
+Route::post('admin/garden/{id}', 'GardenController@storeReview')->name('gardens.review')->middleware('auth');
+Route::delete('admin/garden/{id}', 'GardenController@destroyReview')->name('gardens.review.destroy')->middleware('auth');
 
 Route::resource('garden', 'GardenController');
 Route::get('garden/{id}/reviews', 'GardenController@reviews');
 Route::get('garden/{id}/photos', 'GardenController@photos');
 Route::get('garden/{id}/products', 'GardenController@products');
-Route::post('garden/{id}/follow', 'GardenController@follow');
-Route::post('garden/{id}/unfollow', 'GardenController@unfollow');
+Route::post('garden/{id}/follow', 'GardenController@follow')->middleware('auth');
+Route::post('garden/{id}/unfollow', 'GardenController@unfollow')->middleware('auth');
 Route::get('garden/{id}/estadistica', 'GardenController@estadistica');
 
-Route::resource('admin/garden/{id}/harvest', 'HarvestController');
+Route::resource('admin/garden/{id}/harvest', 'HarvestController')->middleware('auth');
 Route::get('harvest/Photo', 'HarvestController@photo')->name('harvest.photo');
